Fix importing channels with non-ascii names
The zip file that Slack exports uses UTF-8 to encode the filenames, but
doesn't set the proper flags for it so Python assumes the filenames were
encoded using `cp437` (as per the spec). This resulted in channel data
not being extracted to the correct directories and therefore not being
picked up to be imported.

To fix this, zipfile extraction was tweaked so that UTF-8 is always used
to decode filenames. Also, a error will now be logged in the case where
a channel to import doesn't have a matching folder on disk to pull data
from.

Fixes #12