Fix importing custom emojis with -'s in their names
Discord doesn't allow custom emojis with `-`'s in their names. To work
around this, the `-` --> `_` substitution in emoji names that was
already done is now applied for custom emojis as well.

A note was also added to the README to document that users will need to
rename the emojis for them to be used.