Support uploading files that don't support range requests
Not all URLs that Slack provides support range requests, making the
`SeekableHTTPStream` object that's responsible for providing a file-like
interface for the URL report itself as being non-seekable.

This is a problem because the `discord.File()` class requires the
file-like objects provided to it to be seekable.

To fix this, the file-like `CachedSeekableHTTPStream` class was added.
This class behaves exactly like a `SeekableHTTPStream` when range
requests are supported, but has a backup strategy to support seeking if
not. To support seeking without range requests, it will cache the data
as it's downloaded into memory at first, and onto disk if the file ends
up being too big. Seeking forwards downloads more data, while seeking
backwards transparently returns data from the cache.