Prevent failed cached streams from raising exceptions when being closed
If the initial request fails, the `self._cache` attribute of the
`CachedSeekableHTTPStream` would not be set. This would cause the
`close()` method to raise an `AttributeError` when it was called as part
of the `io.BufferedIOBase`'s `__del__` method. Fortunately, because the
exception was raised in the destructor, it was ignored, but still logged
a warning.

This commit fixes the issue by ensuring the `_cache` attribute is set on
the object before making the initial request and potentially raising an
exception.